[Refactor] : 폴더정
