+++ app/src/main/java/bootcamp/sparta/disneym/ui/detail/DetailFragment.kt
@@ -1,19 +1,15 @@
 package bootcamp.sparta.disneym.ui.detail
 
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import bootcamp.sparta.disneym.databinding.FragmentDetailBinding
 import bootcamp.sparta.disneym.model.DetailModel
-import bootcamp.sparta.disneym.ui.MainActivity
 import bootcamp.sparta.disneym.util.Util
 import bootcamp.sparta.disneym.ui.viewmodel.MainSharedEventForDetail
 import bootcamp.sparta.disneym.ui.viewmodel.MainSharedViewModel
