+++ app/src/main/java/bootcamp/sparta/disneym/ui/main/ViewPagerAdapter.kt
@@ -1,4 +1,4 @@
-package bootcamp.sparta.disneym.ui
+package bootcamp.sparta.disneym.ui.main
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
@@ -6,6 +6,7 @@ import androidx.viewpager2.adapter.FragmentStateAdapter
 import bootcamp.sparta.disneym.ui.bookmark.BookmarkFragment
 import bootcamp.sparta.disneym.ui.detail.DetailFragment
 import bootcamp.sparta.disneym.ui.home.HomeFragment
+import bootcamp.sparta.disneym.ui.mypage.MyPageFragment
 import bootcamp.sparta.disneym.ui.search.SearchFragment
 
 /**
@@ -18,7 +19,7 @@ class ViewPagerAdapter(activity: FragmentActivity) : FragmentStateAdapter(activi
 	val fragments: List<Fragment>
 
 	init {
-		fragments = listOf(HomeFragment(), SearchFragment(), BookmarkFragment(), DetailFragment())
+		fragments = listOf(HomeFragment(), SearchFragment(), BookmarkFragment(), MyPageFragment())
 	}
 
 	override fun getItemCount(): Int {
